requirements.txt
# Core ML
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
scipy>=1.10.0
joblib>=1.2.0
tqdm>=4.66.1
imbalanced-learn>=0.10.1

# Feature Engineering + NLP
spacy>=3.5.0
textstat>=0.7.1
sentence-transformers>=2.2.2

# Gradient Boosting
xgboost>=1.7.0
lightgbm>=3.3.5

# Deep Learning
tensorflow>=2.14.0
torch>=2.0.0
torchvision>=0.15.0

# Transformers + Datasets
transformers>=4.33.0
datasets>=2.15.0

# Optimization
optuna>=3.5.0
optuna-dashboard>=0.11.0

# Visualization + Explainability
matplotlib>=3.8.0
seaborn>=0.13.0
shap>=0.44.0
streamlit>=1.31.0
plotly>=5.18.0

# Utils + Configs
pyyaml>=6.0.1
rich>=13.7.0